requirements.txt
numpy 
matplotlib
pandas
python-docx
seaborn 
scikit-learn
xlsxwriter